% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aperm.garray}
\alias{aperm.garray}
\title{General array transposition}
\usage{
\method{aperm}{garray}(a, perm = NULL, ...)
}
\arguments{
\item{a}{A generalized array to be transposed.}

\item{perm}{Desired margins after permutation, integer or character.}

\item{...}{Useless, potential arguments inherited from the S3 generic.}
}
\description{
Restore garray attributes that discarded by aperm.default().
Cannot permute between subdimension (means to promote subdimension of
equal length
into regular dim and reduce dim into subdimension), sorry.
}
