% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{as.data.frame.garray}
\alias{as.data.frame.garray}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{garray}(x, row.names = NULL, optional = FALSE,
  col.names = NULL, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{A generalized array object.}

\item{row.names, optional, stringsAsFactors, ...}{See the same arguments in
?as.data.frame.}

\item{col.names}{'NULL' or a character vector giving the column names.}
}
\description{
Convert a 2D generalized array into a data.frame,
making \code{print()} work correctly.
}
