% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{margins}
\alias{margins}
\alias{margins<-}
\alias{remargins}
\title{The margins and dimensions of a generalized array object}
\usage{
margins(x)

margins(x) <- value

remargins(x, value)
}
\arguments{
\item{x}{A generalized array.}

\item{value}{A character vector will become the margins (names of dimnames)
of the generalized array.
\code{margins<-} ignores the names of \code{value} while
\code{remargins} according to the names of \code{value} renames the margins.
For \code{remargins} the length of value can be shorter than that of the
margins if the value has names.}
}
\description{
Margins means the names of dimnames of an array.
\code{margins<-} and \code{remargins} are for renaming, but
\code{margins<-} ignores the names of \code{value} while
\code{remargins} according to the names of \code{value} renames the margins. Doing
so, \code{remargins} may also keep sdim.  \code{margins<-} always removes sdim.
For \code{remargins} the length of value can be shorter than that of the
margins if the value has names.
}
