% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.garray}
\alias{print.garray}
\title{Print Values}
\usage{
\method{print}{garray}(x, ...)
}
\arguments{
\item{x}{A generalized array object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
Print out a generalized array and returns it \emph{invisibly}.
}
