% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sdim}
\alias{sdim}
\alias{sdim<-}
\title{Subdimensions of an array}
\usage{
sdim(x)

sdim(x, warn = TRUE) <- value
}
\arguments{
\item{x}{A generalized array.}

\item{warn}{Whether issue warning when some of subdimensions are invalid
and get dropped.}

\item{value}{A named list of numeric vectors indicating the
subdivision of some of the dimensions.  The value vill become,
after validated, the attribute sdim.  See '?garray'.}
}
\value{
The subdimensions (a non-empty list) or NULL
}
\description{
Retireve or set the subdimension of an array.
}
\details{
Validation of subdimension is expensive because its consistency with dim need
checking.  Thus most of functions do not validate it.
Operations of subdimension with functions discussed here are guaranteed to
be always keeping the consistency.
}
